% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bench_log.R
\name{extract_bench_log}
\alias{extract_bench_log}
\title{Extract bench-log entries of a certain type}
\usage{
extract_bench_log(
  logs,
  what = c("uptime", "cp_file_source_to_ram", "cp_file_ram_to_drive",
    "cp_file_drive_to_ram", "rm_file_drive", "untar_ram_to_drive", "ls_recursive_drive",
    "find_drive", "du_drive", "chmod_recursive_drive", "tar_drive_to_ram",
    "tar_drive_to_drive", "gzip_drive_to_drive", "rm_folder_drive", "total_time"),
  parse = TRUE
)
}
\arguments{
\item{logs}{A bench_log data.frame}

\item{what}{Which type of entries to extract}

\item{parse}{If TRUE, the `command` column of the extract entries
is parsed and the parsed valus are appended as new columns.
This applies only to `uptime` and `total_time`.}
}
\value{
A [tibble::tibble] data.frame
}
\description{
Extract bench-log entries of a certain type
}
\examples{
path <- system.file(package = "wyntonbench", "exdata", mustWork = TRUE)

## Read *all* raw bench log files
raw <- read_all_bench_logs(path)
trimmed <- trim_bench_log(raw)
print(trimmed)

## Extract a few types of stats
stats <- extract_bench_log(trimmed, what = "cp_file_source_to_ram")
print(stats)

stats <- extract_bench_log(trimmed, what = "untar_ram_to_drive")
print(stats)
}
