% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bench_logs.R
\name{import_bench_logs}
\alias{import_bench_logs}
\title{Import raw bench log files and save per hostname}
\usage{
import_bench_logs(root = "wynton-bench-logs", skip = TRUE,
  verbose = FALSE)
}
\arguments{
\item{root}{(character) A folder to be imported.}

\item{skip}{(logical) Should already existing output files be skipped,
or be recreated.}

\item{verbose}{(logical) Display verbose messages or not.}
}
\value{
A character vector of \file{*.bench_log.rds} files.
}
\description{
Import raw bench log files and save per hostname
}
