% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_bench_log.R
\name{read_all_bench_logs}
\alias{read_all_bench_logs}
\title{Read multiple bench log files at once}
\usage{
read_all_bench_logs(path, recursive = TRUE, pattern = "[.]log$", ...)
}
\arguments{
\item{path}{The path to bench log files.}

\item{recursive}{If TRUE, bench log files are searched recursively.}

\item{pattern}{A regular expression matching files to be read.}

\item{\dots}{Addition arguments passed to \link{read_bench_log}.}
}
\value{
A tibble.
}
\description{
Read multiple bench log files at once
}
\examples{
path <- system.file("exdata", package="wyntonbench")
data <- read_all_bench_logs(path)
print(data)

}
