% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bench_log.R
\name{read_bench_log}
\alias{read_bench_log}
\title{Read bench log files}
\usage{
read_bench_log(file, n_max = +Inf)
}
\arguments{
\item{file}{The bench log file.}

\item{n_max}{Maximum number of lines to read.}
}
\value{
A tibble.
}
\description{
Read bench log files
}
\examples{
file <- system.file("exdata", "ex-bench.log", package="wyntonbench")
data <- read_bench_log(file)
print(data)

}
