% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_bench_log.R
\name{split_bench_log}
\alias{split_bench_log}
\title{Split up a bench log by entry type and saved to individual files}
\usage{
split_bench_log(
  logs,
  whats = eval(formals(extract_bench_log)$what),
  prefix = NULL,
  parse = TRUE,
  skip = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{logs}{A bench_log data.frame}

\item{whats}{(character) Type of entries to extract}

\item{prefix}{(character) The prefix to use for each file written}

\item{parse}{If TRUE, the extracted entries are parsed by
\link{extract_bench_log}.}

\item{skip}{(logical) Should already existing output files be skipped,
or be recreated.}

\item{verbose}{(logical) Display verbose messages or not.}
}
\value{
A character vector of \file{*.bench_log.rds} files.
}
\description{
Split up a bench log by entry type and saved to individual files
}
