% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_bench_log.R
\name{trim_bench_log}
\alias{trim_bench_log}
\title{Drop non-informative entries}
\usage{
trim_bench_log(data)
}
\arguments{
\item{data}{A bench_log data.frame}
}
\value{
A \link[tibble:tibble]{tibble::tibble} data.frame of class \code{bench_log}.
}
\description{
Drop non-informative entries
}
