% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/999.package.R
\docType{package}
\name{wyntonbench-package}
\alias{wyntonbench-package}
\alias{wyntonbench}
\title{Package wyntonbench}
\description{
Tools for benchmarking the Wynton HPC environment.
Some useful functions:
}
\details{
\itemize{
\item \code{\link[=extract_bench_log]{extract_bench_log()}}
}
}
\keyword{package}
