% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bench_log.R
\name{extract_bench_log}
\alias{extract_bench_log}
\title{Extract bench-log entries of a certain type}
\usage{
extract_bench_log(logs, what = c("uptime", "cp-file-source-to-ram",
  "cp-file-ram-to-drive", "cp-file-drive-to-ram", "rm-file-drive",
  "untar-ram-to-drive", "ls-recursive-drive", "find-drive", "du-drive",
  "chmod-recursive--drive", "tar-drive-to-ram", "tar-drive-to-drive",
  "gzip-drive-to-drive", "rm-folder-drive", "total-time"), parse = TRUE)
}
\arguments{
\item{logs}{A bench_log data.frame}

\item{what}{Which type of entries to extract}

\item{parse}{If TRUE, the `command` column of the extract entries
is parsed and the parsed valus are appended as new columns.
This applies only to `uptime` and `total-time`.}
}
\value{
A [tibble::tibble] data.frame
}
\description{
Extract bench-log entries of a certain type
}
